+++ core/src/main/java/com/financeiramente/core/repository/FaturaRepository.java
@@ -2,6 +2,7 @@ package com.financeiramente.core.repository;
 
 import com.financeiramente.core.domain.entity.Fatura;
 import com.financeiramente.core.domain.vo.StatusFatura;
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -13,6 +14,6 @@ public interface FaturaRepository {
     List<Fatura> listarPorCartao(String cartaoId);
     List<Fatura> listarAbertas();
     boolean existeFaturaComStatus(String cartaoId, StatusFatura status);
-    double somarTotalUtilizadoPorCartao(String cartaoId);
-    double somarValorPagoPorDataDePagamento(int ano, int mes);
+    BigDecimal somarTotalUtilizadoPorCartao(String cartaoId);
+    BigDecimal somarValorPagoPorDataDePagamento(int ano, int mes);
 }
