+++ core/src/main/java/com/financeiramente/core/usecase/cartao/EditarCartaoCreditoInput.java
@@ -1,19 +1,20 @@
 package com.financeiramente.core.usecase.cartao;
 
 import com.financeiramente.core.domain.vo.BandeiraCartao;
+import java.math.BigDecimal;
 
 public class EditarCartaoCreditoInput {
     private final String id;
     private final String nome;
     private final int diaVencimento;
     private final int diasParaFechamento;
-    private final Double limite;
+    private final BigDecimal limite;
     private final BandeiraCartao bandeira;
     private final String icone;
     private final String cor;
 
     public EditarCartaoCreditoInput(String id, String nome, int diaVencimento,
-                                    int diasParaFechamento, Double limite,
+                                    int diasParaFechamento, BigDecimal limite,
                                     BandeiraCartao bandeira, String icone, String cor) {
         this.id = id;
         this.nome = nome;
@@ -29,7 +30,7 @@ public class EditarCartaoCreditoInput {
     public String getNome() { return nome; }
     public int getDiaVencimento() { return diaVencimento; }
     public int getDiasParaFechamento() { return diasParaFechamento; }
-    public Double getLimite() { return limite; }
+    public BigDecimal getLimite() { return limite; }
     public BandeiraCartao getBandeira() { return bandeira; }
     public String getIcone() { return icone; }
     public String getCor() { return cor; }
